sclends_dedupe
Most of this commit includes the work on the dedupe script for
evergreen. These changes represent a *mostly finished product. The
script works and the logging is still turned on for debugging.

I also updated the DBhandler.pm to take a database port which means that
most of the other scripts needed to be changed to take that from the
configuration file.